+++ src/params/currencies/DeleteCurrencyParams.ts
@@ -1,5 +1,5 @@
 import {LightrailResponse} from "../LightrailResponse";
 import {Success} from "../../model/Success";
 
-export interface DeleteCurrencyRequest extends LightrailResponse<Success> {
+export interface DeleteCurrencyResponse extends LightrailResponse<Success> {
 }