+++ src/params/transactions/CapturePendingParams.ts
@@ -1,0 +1,9 @@
+import {LightrailResponse} from "../LightrailResponse";
+import {Transaction} from "../../model";
+
+export interface CapturePendingParams {
+    id: string;
+}
+
+export interface CapturePendingResponse extends LightrailResponse<Transaction> {
+}