+++ src/transactions.test.ts
@@ -2,15 +2,14 @@ import * as chai from "chai";
 import * as Lightrail from "./index";
 import * as uuid from "uuid";
 import {getTransaction, listTransactions} from "./transactions";
+import {LightrailTransactionStep} from "./model";
 
 describe("transactions", () => {
-    const sourceValueId = "transferSoruceValueTest";
-    const valueId = "transactionTestValueId";
+    const sourceValueId = "transferSoruceValueTest1";
+    const valueId = "transactionTestValueId1";
 
     before(async () => {
         before(() => {
-            chai.assert.isString(process.env.LIGHTRAIL_API_PATH, "env var LIGHTRAIL_API_PATH must be set ot run the tests (for example set it in the .env file)");
-            chai.assert.isString(process.env.LIGHTRAIL_API_KEY, "env var LIGHTRAIL_API_KEY must be set ot run the tests (for example set it in the .env file)");
             Lightrail.configure({
                 restRoot: process.env.LIGHTRAIL_API_PATH || "",
                 apiKey: process.env.LIGHTRAIL_API_KEY || "",
@@ -19,22 +18,24 @@ describe("transactions", () => {
 
         const value = await Lightrail.values.getValue(valueId);
 
-        if (!value) {
+        if (!value.body) {
             await Lightrail.values.createValue({
                 id: valueId,
-                code: "TRANSACTION_TEST_CODE",
+                code: "TRANSACTION_TEST_CODE_1",
                 currency: "USD",
+                usesRemaining: 100,
                 balance: 0 // must be 0 for totals test
             });
         }
 
         const srcValue = await Lightrail.values.getValue(sourceValueId);
 
-        if (!srcValue) {
+        if (!srcValue.body) {
             await Lightrail.values.createValue({
                 id: sourceValueId,
-                code: "TRANSACTION_SOURCE_CODE",
+                code: "TRANSACTION_SOURCE_CODE_1",
                 currency: "USD",
+                usesRemaining: 100,
                 balance: 0
             });
         }
@@ -48,6 +49,21 @@ describe("transactions", () => {
                 id: creditId,
                 currency: "USD",
                 amount: 100000,
+                destination: {
+                    rail: "lightrail",
+                    valueId: valueId
+                }
+            });
+
+            chai.assert.isNotNull(credit);
+            chai.assert.equal(credit.body.transactionType, "credit");
+        });
+
+        it("successfully creates a credit transaction with uses instead of amount", async () => {
+            const credit = await Lightrail.transactions.credit({
+                id: uuid.v4().substring(0, 24),
+                currency: "USD",
+                uses: 2,
                 destination: {
                     rail: "lightrail",
                     valueId: valueId
@@ -74,7 +90,7 @@ describe("transactions", () => {
                 sources: [
                     {
                         rail: "lightrail",
-                        code: "TRANSACTION_TEST_CODE"
+                        code: "TRANSACTION_TEST_CODE_1"
                     }
                 ]
             });
@@ -95,7 +111,7 @@ describe("transactions", () => {
         });
     });
 
-    describe("transfer", () => {
+    describe("transfer()", () => {
         it("successfully creates a transfer transaction", async () => {
             const updateSource = await Lightrail.transactions.credit({
                 id: uuid.v4().substring(0, 24),
@@ -127,7 +143,7 @@ describe("transactions", () => {
     });
 
     describe("debit()", () => {
-        it("successfully creates a credit transaction", async () => {
+        it("successfully creates a debit transaction", async () => {
             const transaction = await Lightrail.transactions.debit({
                 id: uuid.v4().substring(0, 24),
                 currency: "USD",
@@ -140,6 +156,101 @@ describe("transactions", () => {
 
             chai.assert.isNotNull(transaction);
             chai.assert.equal(transaction.body.transactionType, "debit");
+        });
+
+        it("successfully creates a debit transaction with uses instead of amount", async () => {
+            const transaction = await Lightrail.transactions.debit({
+                id: uuid.v4().substring(0, 24),
+                currency: "USD",
+                uses: 2,
+                source: {
+                    rail: "lightrail",
+                    valueId: valueId
+                }
+            });
+
+            chai.assert.isNotNull(transaction);
+            chai.assert.equal(transaction.body.transactionType, "debit");
+        });
+    });
+
+    describe("reverse()", () => {
+        it("successfully reverses a transaction", async () => {
+            const creditTx = await Lightrail.transactions.credit({
+                id: uuid.v4().substring(0, 24),
+                currency: "USD",
+                amount: 549,
+                destination: {
+                    rail: "lightrail",
+                    valueId: valueId
+                }
+            });
+
+            chai.assert.isNotNull(creditTx);
+            chai.assert.equal(creditTx.body.transactionType, "credit");
+
+            const reverseTx = await Lightrail.transactions.reverse(creditTx.body, {
+                id: uuid.v4().substring(0, 24)
+            });
+
+            chai.assert.isNotNull(reverseTx);
+            chai.assert.equal(reverseTx.body.transactionType, "reverse");
+            chai.assert.equal((reverseTx.body.steps[0] as LightrailTransactionStep).balanceAfter, (creditTx.body.steps[0] as LightrailTransactionStep).balanceBefore);
+            chai.assert.equal((reverseTx.body.steps[0] as LightrailTransactionStep).balanceBefore, (creditTx.body.steps[0] as LightrailTransactionStep).balanceAfter);
+            chai.assert.equal((reverseTx.body.steps[0] as LightrailTransactionStep).balanceChange, -(creditTx.body.steps[0] as LightrailTransactionStep).balanceChange);
+        });
+    });
+
+    describe("pendingVoid()", () => {
+        it("successfully voids a pending transaction", async () => {
+            const debitTx = await Lightrail.transactions.debit({
+                id: uuid.v4().substring(0, 24),
+                currency: "USD",
+                amount: 100,
+                source: {
+                    rail: "lightrail",
+                    valueId: valueId
+                },
+                pending: true
+            });
+
+            chai.assert.isNotNull(debitTx);
+            chai.assert.equal(debitTx.body.transactionType, "debit");
+
+            const voidTx = await Lightrail.transactions.voidPending(debitTx.body, {
+                id: uuid.v4().substring(0, 24)
+            });
+
+            chai.assert.isNotNull(voidTx);
+            chai.assert.equal(voidTx.body.transactionType, "void");
+            chai.assert.equal((voidTx.body.steps[0] as LightrailTransactionStep).balanceAfter, (debitTx.body.steps[0] as LightrailTransactionStep).balanceBefore);
+            chai.assert.equal((voidTx.body.steps[0] as LightrailTransactionStep).balanceBefore, (debitTx.body.steps[0] as LightrailTransactionStep).balanceAfter);
+            chai.assert.equal((voidTx.body.steps[0] as LightrailTransactionStep).balanceChange, -(debitTx.body.steps[0] as LightrailTransactionStep).balanceChange);
+        });
+    });
+
+    describe("pendingCapture()", () => {
+        it("successfully captures a pending transaction", async () => {
+            const debitTx = await Lightrail.transactions.debit({
+                id: uuid.v4().substring(0, 24),
+                currency: "USD",
+                amount: 100,
+                source: {
+                    rail: "lightrail",
+                    valueId: valueId
+                },
+                pending: true
+            });
+
+            chai.assert.isNotNull(debitTx);
+            chai.assert.equal(debitTx.body.transactionType, "debit");
+
+            const captureTx = await Lightrail.transactions.capturePending(debitTx.body, {
+                id: uuid.v4().substring(0, 24)
+            });
+
+            chai.assert.isNotNull(captureTx);
+            chai.assert.equal(captureTx.body.transactionType, "capture");
         });
     });